Get LocalCI running on as many Rubies as easily possible
Basically I am just using ones that have preexisting Docker images. I
didn't do JRuby ARM64 because it doesn't have `git` and that was causing
the `gemspec` file to fail to run.

Bundler was locked to 2.6.9 because of a fairly silly reason. Ruby 2.7
and 3.0 worked fine but 3.1 broke because it has a recent enough version
of bundler that updates itself, and bundler 2.7 doesn't work on Ruby
3.1. I don't love this but it's a small change to not have to have
multiple Gemfiles. Down the line when I put more effort into testing
more Ruby versions I'll just do a Gemfile for each version.